GUI Modifications
Modified the GUI window and callbacks scripts.  Main changes reflect the new data structure organization:
Data Structure - mf_fitter
Main fields
user_inputs - this is for variables input by the user in the various text boxes of the GUI.  It includes the control parameter, intensity cut-off, peak separation cut-off, fitting reference files, and smoothing preferences.
fit_data - this will store all of the fitted parameters after each cycle of the fitting algorithm. While this will vary depending on what function the user has decided to fit, for the most common purposes it will contain the background, intensity, fwhm, and center for two or three peaks along with error from the 90% confidence interval.
algorithm_options - again, this will vary has more algorithms are incorporated.  Currently it stores checkbox selection on which of the fitting cycles the user wants to have displayed vs running in the background.
save_options - this stores the extension / folder string for saving files, as well as what file types (.jpg, .fig, .eps, .pdf) the user wants graphs/figures to be saved as.
handles - this stores all of the handles pointing to the various objects such as the main GUI window and all of the text boxes for user inputs.
numors - stores the data file numbers loaded into GRASP over which mf_fitter is to be run
depth - stores the maximum number of files, this variable seems unnecessary and could be accomplished by replacing "mf_fitter.depth" with length(mf_fitter.numors) everywhere in a future version.

Things that still need to be fixed in these files in a future version:
- add a switch so the sector box can apply to I vs Q and I vs Xi fittings
- and a # of peaks text box or check box next to the reference files instead of assuming it is one or two
- there are some missing algorithm scripts in the fitting algorithm drop down (i.e. Raster Scan)
- open the curve fit control window automatically during intialization
- incorporate mf_fitter_plotMaker.m when that script is complete
- remove any duplicate GRASP functions (i.e. may sector_box_callbacks2) that are unnecessary and lingering from a prior version